Make likelihood function smoothly plotted
Have the likelihood function be plotted by using linspace of 1000 points
to smoothly generate the function's shape. The sliders now only control
the number of data points that are sampled.

In the LaTeX use "\,\middle|\," for the conditional vertical bar instead
of using "\,\big|\,". This will make things work better with "\left("
and "\right)".