Add flow of information for prediction and inference
This is to try and show something similar to Kyle Cranmer's slide that
shows how to use p(x,z | theta, nu) to move between the parameters of
interest and nuisance parameters (prediction) to the observed data, and
likewise how to use the observed data and latent variables (inference)
to move to the parameters of interest and nuisance parameters.

This slide is slide 19 in his Chaire Georges Lemaitre 2017 talk.
c.f.: https://agenda.irmp.ucl.ac.be/conferenceDisplay.py?confId=2658
      https://figshare.com/articles/Rethinking_Physics_in_the_Age_of_Deep_Learning/5639239